+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - update atime for infiles when (re)submit; to avoid purging files prematurely (#1689)
 - outputs.json shall contain relpaths instead of abspaths (#1652)
 - runs' output folders shall once again contain a copy of the WDL, inputs-JSON, and subworkflows-ZIP (#1710)
+- fixed bug when Cromwell submission fails during input processing, before a folder is created, and was not recognized by JAWS (#1711)
 
 ## 2.0.0 Summary [08/23/2023]
 ### Internal Facing Changes